Fix for MapServer WFS Projection Handling Bug(s)
As seen here:
 - https://lists.osgeo.org/pipermail/mapserver-users/2011-December/071117.html

The problem is getting features back in a projection that is not 4326
when the underlaying dataset is not in 4326.  MapServer appears able to
reproject from the actual data projection to 4326, but not to other
projection (including web mercator) when the base data is in something
like UTM 15N.

This "hack" has MapServer WFS layers doing all of their communication
in 4326.