Add explicit build-deploy grunt task.
This is a shortcut to save time while building and testing the
deploy (and not also testing the development build).  I've been
doing this manually for awhile by explicitly listing the tasks
on the grunt command line.  It saves about half the time on a
build.

`grunt build` still builds everything as before (and as needed by
`npm pack`).