Added visibility toggle for the grid.
Adds the ability to min/maximize the grid. The button even
does a fancy icon change.

This also cleans up some of the differing CSS between the toolbar
and the grid.  Longer term, we should initiate a project that
reuses more components but for now, it seemed like a good idea
to at least fix the CSS better.