Use 3857 coordinates in PDF.
This reverts f377e1 and bc4d96 and replaces the proj string in the
PDF with one for 3857 (that works in the viewers).  This avoids all
the geometry issues we were having.

TODO: Support for other than 3857 maps will be difficult because of
      the specifically formatted embedded projection string.
      Realistically the admin will have to supply this somehow.

TODO: The coordinates scale is wrong when printed at higher or highest
      resolutions.