Added a control to allow clearing selections
After all queries have been removed from the selection tab,
the selection polygon would continue to be displayed.

This new functionality fixes that issue by adding a button
allowing the user to clear the polygon.  This was done as a bit
of a compromise. Ostensibly, a complex selection area may need to be
re-used even if the user wants to clear teh current results.  Automatically
removing the selection area would've made that impossible.

Steps to use:
1. run a query with a drawn selection feature.
2. click "delete" / trash-can icon on the results.
3. See that there is now a polygon on display.
4. click the 'clear' button in the service manager tab.