Fix issue with disappearing tool interaction
When "End Drawing" is clicked it sets the interactionType to null.
However, when re-clicking the same service, this was causing a weird
null-state to be displayed to the user. They had started the service
but no drawing tool would be selected.

This restores the drawing tool to the default no matter what.